refactor ui/view package
