+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
         android:theme="@style/AppTheme">
 
         <activity
-            android:name=".ui.MainActivity"
+            android:name=".view.MainActivity"
             android:label="@string/app_name"
             android:foregroundServiceType="location"
             android:theme="@style/AppTheme.NoActionBar">
@@ -31,8 +31,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ui.PlaceSearchActivity"
-            android:parentActivityName=".ui.MainActivity">
+        <activity android:name=".view.PlaceSearchActivity"
+            android:parentActivityName=".view.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.SEARCH" />
             </intent-filter>
